# public/swagger.json
{
	"components": {
...
		"schemas": {
			"Response": {
				"properties": {
					"data": {},
					"status": {
						"type": "number",
						"format": "double"
					},
					"error": {
						"type": "string"
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"data",
					"status",
					"error",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"saleInvoice": {
				"properties": {
					"billedFrom": {
						"type": "string"
					},
					"billedTo": {
						"type": "string"
					},
					"shippingAddress": {
						"type": "string"
					},
					"invoiceDate": {
						"type": "string"
					},
					"dispatchThrough": {
						"type": "string"
					},
					"products": {
						"items": {
							"properties": {
								"discount": {
									"type": "number",
									"format": "double"
								},
								"rate": {
									"type": "number",
									"format": "double"
								},
								"quantity": {
									"type": "number",
									"format": "double"
								},
								"productId": {
									"type": "string"
								}
							},
							"required": [
								"discount",
								"rate",
								"quantity",
								"productId"
							],
							"type": "object"
						},
						"type": "array"
					},
					"totalAmount": {
						"type": "number",
						"format": "double"
					},
					"isBlacked": {
						"type": "boolean"
					},
					"id": {
						"type": "string"
					}
				},
				"required": [
					"billedFrom",
					"billedTo",
					"shippingAddress",
					"invoiceDate",
					"dispatchThrough",
					"products",
					"totalAmount"
				],
				"type": "object",
				"additionalProperties": false
			},
			"saleInvoiceSerial": {
				"properties": {
					"saleInvoiceId": {
						"type": "string"
					},
					"itemId": {
						"type": "string"
					},
					"serialNumber": {
						"type": "string"
					}
				},
				"required": [
					"saleInvoiceId",
					"itemId",
					"serialNumber"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"Bearer": {
				"type": "apiKey",
				"name": "Authorization",
				"in": "header",
				"description": "JWT Authorization header using the Bearer scheme."
			}
		}
	},
	"info": {
		"title": "inventory-system",
		"version": "0.1.3",
		"description": "Inventory Management System",
		"license": {
			"name": "license"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/invoice/sale/save": {
			"post": {
				"operationId": "Save",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Response"
								}
							}
						}
					}
				},
				"tags": [
					"Invoice/Sale"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/saleInvoice"
							}
						}
					}
				}
			}
		},
		"/invoice/sale/getAll": {
			"get": {
				"operationId": "GetAll",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Response"
								}
							}
						}
					}
				},
				"tags": [
					"Invoice/Sale"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "pageNumber",
						"required": false,
						"schema": {
							"default": 1,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "pageSize",
						"required": false,
						"schema": {
							"default": 20,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "status",
						"required": false,
						"schema": {
							"default": "",
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "isBlacked",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/invoice/sale/get": {
			"get": {
				"operationId": "Get",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Response"
								}
							}
						}
					}
				},
				"tags": [
					"Invoice/Sale"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/invoice/sale/serialNumberEntry": {
			"post": {
				"operationId": "SerialNumberEntry",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Response"
								}
							}
						}
					}
				},
				"tags": [
					"Invoice/Sale"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/saleInvoiceSerial"
							}
						}
					}
				}
			}
		},
		"/invoice/sale/confirmInvoice": {
			"post": {
				"operationId": "ConfirmInvoice",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Response"
								}
							}
						}
					}
				},
				"tags": [
					"Invoice/Sale"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"saleInvoiceId": {
										"type": "string"
									}
								},
								"required": [
									"saleInvoiceId"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/invoice/sale/approveInvoice": {
			"post": {
				"operationId": "ApproveInvoice",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Response"
								}
							}
						}
					}
				},
				"tags": [
					"Invoice/Sale"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"saleInvoiceId": {
										"type": "string"
									}
								},
								"required": [
									"saleInvoiceId"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/invoice/sale/updateSerialNumber": {
			"post": {
				"operationId": "UpdateSerialNumber",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Response"
								}
							}
						}
					}
				},
				"tags": [
					"Invoice/Sale"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"newSerialNumber": {
										"type": "string"
									},
									"oldSerialNumber": {
										"type": "string"
									}
								},
								"required": [
									"newSerialNumber",
									"oldSerialNumber"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/invoice/sale/cancelSaleInvoice": {
			"patch": {
				"operationId": "CancelSaleInvoice",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Response"
								}
							}
						}
					}
				},
				"tags": [
					"Invoice/Sale"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"cancel": {
										"type": "boolean"
									},
									"id": {
										"type": "string"
									}
								},
								"required": [
									"cancel",
									"id"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/invoice/sale/delete": {
			"delete": {
				"operationId": "Delete",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Response"
								}
							}
						}
					}
				},
				"tags": [
					"Invoice/Sale"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}